pyproject.toml
[tool.poetry]
name = "flowapi"
version = "0.1.0"
description = "An Api to interact with the Flow Hololens software from Holoforge"
authors = ["Edouard Leroy <edouard.leroy@cea.fr>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pypugjs = "^5.9.12"
Jinja2 = "^3.1.3"
pillow = "^10.3.0"
pdf2image = "^1.17.0"
xmltodict = "^0.13.0"
weasyprint = "^61.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
flow_ui_generator = 'flowapi.ui_component_generator:ui_generator_command_line'